scripts/requirements.txt
annotated-types==0.7.0
babel==2.17.0
dsw-models==4.26.1
pydantic==2.12.5
pydantic_core==2.41.5
typing-inspection==0.4.2
typing_extensions==4.15.0
